hard-code APK's minSdkVersion, targetSdkVersion is different
ApplicationInfo.targetSdkVersion provides the "minimum target version",
which is the version that the build was targeted to.  This is different
than the minSdkVersion, which is the minimum version that the APK will run
on.  It seems that the only way to get the minSdkVersion from the APK is to
parse it out of the AndroidManifest.xml... sigh...

refs #3016