ActionMode for local repo config
Now the local repo config feels a lot more Android-ish, with the ActionMode
being triggering when the user makes a change to the list of selected apps.
The ProgressDialog for generating the repo is now shown on top of the
KerplappActivity, and it is cancelable via touch outside or back button,
and the repo generation process will keep going in the background.

fixes #3200 https://dev.guardianproject.info/issues/3200