move webserver to a service for hosting all swapping stuff
The webserver thread is now in a separate Service that is meant to host all
the stuff related to swapping repos and apps.  Anything related to swapping
that needs to occur in the background so that the user can do whatever with
FDroid should go in this service.  That then provides an easy way to start
and stop swapping as well as provide a notification for when swapping is
active.

fixes #2908 https://dev.guardianproject.info/issues/2908