create RESTART msg and wire it up to LocalRepoService
This is needed to restart the webserver when the wifi or IP address changes
so that it can bind to the new address.  I added a STOP and START message
for good measure, they will be useful for controlling the webserver based
on the status of swapping.

fixes #2908 https://dev.guardianproject.info/issues/2908