only trigger WifiStateChangeService when wifi is enabling or enabled
WifiStateChangeService waits for WiFi to be enabled and set up with an IP
address before it will continue.  Therefore, only trigger it when WiFi is
already enabled or in the process of becoming so.  Otherwise, no need to
worry since WifiStateChangeReceiver will trigger whenever wifi is connected.

This also adds a new trigger to the "Use HTTPS" preference, so that it will
take effect immediately when the checkbox is changed.