refactor webserver class into much simplified NanoHTTPD subclass
The SimpleWebServer.java example had a lot more stuff than we needed in it,
plus it had some code that was only supported on android-9 or higher. I
stripped out anything that is not currently needed and merged
KerplappHTTPD and SimpleWebServer into a single class: LocalHTTPD.

This should now work down to android-7 at least.

refs #3204 https://dev.guardianproject.info/issues/3204